Recursive CNN
It is the recursive version of Convolutional Neural Network(CNN) which is designed to make CNN deeper in order to increase the accuracy of image classification. In this way to have better accuracy it is essential to adjust different parameters of this recursive loop to measure the accuracy of classification for various amount of these parameters.